requirements.txt
# Core
Django>=5.2,<7
djangorestframework>=3.15
django-filter>=24.3
djangorestframework-simplejwt[crypto]>=5.5
django-cors-headers>=4.6

# Database
psycopg2-binary>=2.9.9

# Configuration
python-dotenv>=1.0.1
dj-database-url>=2.1.0

# Payments
stripe>=11,<13

# Media / future image handling
Pillow>=10.0
